move addTooltip to ToolBarButton
