+++ lib/items/tool-bar-button-view.js
@@ -1,5 +1,5 @@
 const {CompositeDisposable} = require('atom');
-const { ToolBarItem, addTooltip } = require('./tool-bar-item');
+const { ToolBarItem } = require('./tool-bar-item');
 
 /**
  * A button class with many options
@@ -46,7 +46,7 @@ class ToolBarButtonView extends ToolBarItem {
     if (options.tooltip) {
       this.subscriptions = new CompositeDisposable();
       this.subscriptions.add(
-        addTooltip(this.options.tooltip, this.options.callback)
+        this.addTooltip(this.options.tooltip, this.options.callback)
       );
     }
 
@@ -97,6 +97,35 @@ class ToolBarButtonView extends ToolBarItem {
     } else {
       this.element.textContent = this.options.text;
     }
+  }
+
+  /**
+   * adds a Tooltip for your item.
+   * @param {ButtonOptions.tooltip} tooltipOptions
+   * @param {ButtonOptions.callback | null} callback
+   * @returns {Disposable} a disposable tooltip
+   */
+  addTooltip (tooltipOptions, callback = null) {
+    let tooltip;
+    if (typeof tooltipOptions === 'string') {
+      tooltip = {
+        title: tooltipOptions
+      };
+    } else {
+      tooltip = tooltipOptions;
+    }
+
+    if (!tooltip.hasOwnProperty('placement')) {
+      tooltip.placement = getTooltipPlacement();
+    }
+
+    if (!tooltip.hasOwnProperty('keyBindingCommand') &&
+      typeof callback === 'string'
+    ) {
+      tooltip.keyBindingCommand = callback;
+    }
+
+    return atom.tooltips.add(this.element, tooltip);
   }
 
   /** Add color to the button */
@@ -216,6 +245,17 @@ function getCallbackModifier (callback, {altKey, ctrlKey, shiftKey}) {
       return true;
     });
   return callback[modifier] || callback[''];
+}
+
+/** get the tooltip placement based on the toolbar position */
+function getTooltipPlacement () {
+  const toolbarPosition = atom.config.get('tool-bar.position');
+
+  return toolbarPosition === 'Top' ? 'bottom'
+    : toolbarPosition === 'Right' ? 'left'
+      : toolbarPosition === 'Bottom' ? 'top'
+        : toolbarPosition === 'Left' ? 'right'
+          : null;
 }
 
 module.exports.ToolBarButtonView = ToolBarButtonView;
