+++ lib/items/tool-bar-item.js
@@ -26,45 +26,3 @@ class ToolBarItem {
 }
 
 module.exports.ToolBarItem = ToolBarItem;
-
-/**
- * adds a Tooltip for your item.
- * @param {ButtonOptions.tooltip} tooltipOptions
- * @param {ButtonOptions.callback | null} callback
- * @returns {Disposable} a disposable tooltip
- */
-function addTooltip (tooltipOptions, callback = null) {
-  let tooltip;
-  if (typeof tooltipOptions === 'string') {
-    tooltip = {
-      title: tooltipOptions
-    };
-  } else {
-    tooltip = tooltipOptions;
-  }
-
-  if (!tooltip.hasOwnProperty('placement')) {
-    tooltip.placement = getTooltipPlacement();
-  }
-
-  if (!tooltip.hasOwnProperty('keyBindingCommand') &&
-      typeof callback === 'string'
-  ) {
-    tooltip.keyBindingCommand = callback;
-  }
-
-  return atom.tooltips.add(this.element, tooltip);
-}
-
-/** get the tooltip placement based on the toolbar position */
-function getTooltipPlacement () {
-  const toolbarPosition = atom.config.get('tool-bar.position');
-
-  return toolbarPosition === 'Top' ? 'bottom'
-    : toolbarPosition === 'Right' ? 'left'
-      : toolbarPosition === 'Bottom' ? 'top'
-        : toolbarPosition === 'Left' ? 'right'
-          : null;
-}
-
-module.exports.addTooltip = addTooltip;
